0417정/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-seed==0.3.1
djangorestframework==3.14.0
Faker==18.4.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
toposort==1.10
